Improve the fem pipeline test Add petsc generated c code to .gitignore
